Start by adding the idea of a transform and a pipeline into the playground.
Start by adding the idea of a transform and a pipeline into the playground. The idea here is to be able to quickly play around with various image transforms via a command line interface, by creating a set of configurable transforms, and then selectively being able to apply different combinations of them dynamically, and see the results without a recompile step.